serving fonts statically, factoring utilities of search component
