+++ client/src/css/app.css
@@ -1,20 +1,17 @@
 @font-face {
   font-family: 'AdiNeuePro-Black';
-  src: url('https://www.adidas.com/glass/assets/fonts/adidas/adineuePRO-Black.woff2') format('woff2');
+  src: url('/assets/fonts/adineuePRO-Black.woff2') format('woff2');
 }
-
 
 @font-face {
   font-family: 'AdiNeuePro';
-  src: url('https://www.adidas.com/glass/assets/fonts/adidas/adineuePRO-Regular.woff2?hash=1RLZb') format('woff2');
+  src: url('/assets/fonts/adineuePRO-Regular.woff2') format('woff2');
 }
 
 @font-face {
   font-family: "AdihausDIN";
-  src: url("https://www.adidas.com/glass/assets/fonts/adidas/AdihausDIN-Regular.woff2") format('woff2');
+  src: url("/assets/fonts/AdihausDIN-Regular.woff2") format('woff2');
 }
-
-
 
 * { box-sizing: border-box }
